Food Inventory
Added in a very simple food inventory UI. I have the ability now to add
the name and place it in an unordered list. I will extend the
functionality furhter once I have some more tests in place.

I like the Outside-In testing methodology. I have used it before and
have called it "Get end to end as fast as possible". Outside-In also is
an interesting concept.